requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
asyncpg==0.29.0
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9
celery==5.3.6
redis==5.0.1
psycopg2-binary==2.9.9
httpx==0.27.0
# For local LLM integration later
# vllm
# openai
# Audio
faster-whisper==0.10.0
# Utilities
requests==2.31.0
pyjwt==2.8.0
python-dateutil==2.8.2
# Charts & Market Data
matplotlib==3.8.3
yfinance==0.2.36
